* added public html folder * added local nodes directory with sample node (not active), but speed ups node development if necessary * moved source to ./src directory
