+++ .eslintrc.json
@@ -9,10 +9,6 @@
     "linebreak-style": [
       "error",
       "unix"
-    ],
-    "quotes": [
-      "warn",
-      "single"
     ]
   }
 }
