+++ settings.js
@@ -144,9 +144,9 @@ module.exports = {
       level: "debug",
       metrics: false,
       audit: false,
-      handler: require("./logger_con.js")
-    }
-    /*
+      handler: require("./src/logger_con.js")
+    },
+    
     // file logger
     file: {
       level: "info",
@@ -155,9 +155,9 @@ module.exports = {
       logfilename: path.join(process.cwd(), "nr.log"), // the default logfile
       divider: "\t", // col divider
       newline: "\n", // newline characters typically \r\n (CRLF) or \n (LF)
-      handler: require("./logger_file.js")
+      handler: require("./src/logger_file.js")
     },
-    */
+ 
     /*
     // syslog
     syslog: {
@@ -167,7 +167,7 @@ module.exports = {
       syslogserver: "udp://nas:514/NR", // syslog server [tcp|udp]://host[:port]/program
       rfc3164: true, //set to false to use RFC 5424 syslog header format; default is true for the older RFC 3164 format.
       facility: "Local0", //  Kernel, User, System, Audit, Alert, Local0 - Local7
-      handler: require("./logger_syslog.js")
+      handler: require("./src/logger_syslog.js")
     }
     */
   },
