+++ src/server.js
@@ -1,6 +1,7 @@
 /*eslint-env node */
 /*eslint indent:0,semi:0 */
 /*jshint devel: true, node: true*/
+/* eslint no-console: 0 */
 /*global: */
 
 /***
@@ -27,7 +28,7 @@ const http = use_https ? require("https") : require("http");
 
 const express = require("express"); // THE std library for serving HTTP
 const RED = require("node-red");
-var nrSettings = require("./settings.js"); // Node-Red settings file
+var nrSettings = require("../settings.js"); // Node-Red settings file
 const fs = require("fs");
 
 // you can set a default credential secret for storing node's credentials within node red
@@ -57,15 +58,15 @@ if (process.env.npm_package_config_nr_title) {
 var app = express();
 
 // Add a simple route for static content served from './public'
-app.use("/", express.static("public"));
+app.use("/", express.static("./public"));
 
 // Add static route for bower components from './bower_components'
-app.use("/bower_components", express.static("bower_components"));
+app.use("/bower_components", express.static("./bower_components"));
 
 // Create the http(s) server
 if (use_https) {
-  var privateKey = fs.readFileSync("server.key", "utf8");
-  var certificate = fs.readFileSync("server.crt", "utf8");
+  var privateKey = fs.readFileSync("./server.key", "utf8");
+  var certificate = fs.readFileSync("./server.crt", "utf8");
   var credentials = {
     key: privateKey,
     cert: certificate
